--- a/data/fr.txt
+++ b/data/fr.txt
@@ -1,0 +1,24 @@
+merde
+putain
+con
+connard
+connasse
+salope
+salaud
+bordel
+chiant
+foutre
+enculé
+enfoiré
+fils de pute
+casse‑toi
+ferme ta gueule
+va te faire foutre
+ta gueule
+pute
+couille
+couillon
+bâtard
+gros con
+cul
+trou du cul