--- a/data/ga.txt
+++ b/data/ga.txt
@@ -1,0 +1,16 @@
+cac
+píosa cac
+muc
+feck
+feckin
+bod
+póg mo thóin
+amadán
+eejit
+gobshite
+bastún
+lúdramán
+balbhán
+ainnis ort
+dún do chlab
+damnú ort