… .DS_Store
… .github/.DS_Store
… .github/readme.png
… README.md
… __tests__/chinese.test.js
… __tests__/engine.test.js
… __tests__/english.test.js
… __tests__/french.test.js
… __tests__/irish.test.js
… __tests__/spanish.test.js
… data/ar.txt
… data/zh.txt